resolve handler conflicts
