--- src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/mediation/MediationSuccessCallbackHandler.java
@@ -1,71 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.mediation;
-
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
-import uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.Role;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.Callback;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackHandler;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
-import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
-import uk.gov.hmcts.cmc.domain.models.Claim;
-import uk.gov.hmcts.cmc.domain.utils.FeaturesUtils;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.MEDIATION_SUCCESSFUL;
-import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights.REFERENCE_NUMBER;
-import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.MEDIATION_PILOT_SUCCESS;
-import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.CASEWORKER;
-
-@Service
-public class MediationSuccessCallbackHandler extends CallbackHandler {
-    private static final List<Role> ROLES = Collections.singletonList(CASEWORKER);
-    private static final List<CaseEvent> EVENTS = ImmutableList.of(MEDIATION_SUCCESSFUL);
-
-    private final CaseDetailsConverter caseDetailsConverter;
-
-    private final AppInsights appInsights;
-
-    @Autowired
-    public MediationSuccessCallbackHandler(
-        CaseDetailsConverter caseDetailsConverter,
-        AppInsights appInsights
-    ) {
-        this.caseDetailsConverter = caseDetailsConverter;
-        this.appInsights = appInsights;
-    }
-
-    @Override
-    protected Map<CallbackType, Callback> callbacks() {
-        return ImmutableMap.of(CallbackType.SUBMITTED, this::raiseAppInsightEvent);
-    }
-
-    @Override
-    public List<CaseEvent> handledEvents() {
-        return EVENTS;
-    }
-
-    @Override
-    public List<Role> getSupportedRoles() {
-        return ROLES;
-    }
-
-    private CallbackResponse raiseAppInsightEvent(CallbackParams callbackParams) {
-        CaseDetails caseDetails = callbackParams.getRequest().getCaseDetails();
-        Claim claim = caseDetailsConverter.extractClaim(caseDetails);
-        if (FeaturesUtils.hasMediationPilotFeature(claim)) {
-            appInsights.trackEvent(MEDIATION_PILOT_SUCCESS, REFERENCE_NUMBER, claim.getReferenceNumber());
-        }
-        return SubmittedCallbackResponse.builder().build();
-    }
-}
