--- src/test/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/mediation/MediationSuccessCallbackHandlerTest.java
@@ -1,77 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.mediation;
-
-import com.google.common.collect.ImmutableMap;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
-import uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights;
-import uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
-import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
-import uk.gov.hmcts.cmc.domain.models.Claim;
-import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
-import uk.gov.hmcts.cmc.domain.utils.FeaturesUtils;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-
-import java.util.Collections;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights.REFERENCE_NUMBER;
-import static uk.gov.hmcts.cmc.claimstore.utils.VerificationModeUtils.once;
-
-@RunWith(MockitoJUnitRunner.class)
-public class MediationSuccessCallbackHandlerTest {
-
-    @Mock
-    private CaseDetailsConverter caseDetailsConverter;
-
-    @Mock
-    private AppInsights appInsights;
-
-    private MediationSuccessCallbackHandler mediationSuccessCallbackHandler;
-
-    private static final String AUTHORISATION = "Bearer: aaaa";
-
-    private Claim claimSetForMediation =
-        SampleClaim.getWithClaimantResponseRejectionForPartAdmissionAndMediation();
-
-    @Before
-    public void setUp() {
-        mediationSuccessCallbackHandler = new MediationSuccessCallbackHandler(caseDetailsConverter, appInsights);
-    }
-
-    @Test
-    public void shouldRaiseAppInsight() {
-
-        Claim claim = claimSetForMediation.toBuilder()
-            .features(Collections.singletonList(FeaturesUtils.MEDIATION_PILOT))
-            .build();
-
-        when(caseDetailsConverter.extractClaim(any(CaseDetails.class))).thenReturn(claim);
-
-        CallbackRequest callbackRequest = CallbackRequest
-            .builder()
-            .caseDetails(CaseDetails.builder().data(Collections.emptyMap()).build())
-            .eventId(CaseEvent.MEDIATION_SUCCESSFUL.getValue())
-            .build();
-
-        CallbackParams callbackParams = CallbackParams.builder()
-            .type(CallbackType.SUBMITTED)
-            .request(callbackRequest)
-            .params(ImmutableMap.of(CallbackParams.Params.BEARER_TOKEN, AUTHORISATION))
-            .build();
-
-        mediationSuccessCallbackHandler.handle(callbackParams);
-
-        verify(appInsights, once()).trackEvent(eq(AppInsightsEvent.MEDIATION_PILOT_SUCCESS),
-            eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
-    }
-}
