+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/mediation/MediationSuccessfulCallbackHandlerTest.java
@@ -6,6 +6,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights;
+import uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent;
 import uk.gov.hmcts.cmc.claimstore.config.properties.notifications.EmailTemplates;
 import uk.gov.hmcts.cmc.claimstore.config.properties.notifications.NotificationTemplates;
 import uk.gov.hmcts.cmc.claimstore.config.properties.notifications.NotificationsProperties;
@@ -17,15 +19,18 @@ import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.cmc.domain.models.response.YesNoOption;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleResponse;
+import uk.gov.hmcts.cmc.domain.utils.FeaturesUtils;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 
 import java.util.Collections;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.mockito.ArgumentMatchers.eq;
+import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights.REFERENCE_NUMBER;
+import static uk.gov.hmcts.cmc.claimstore.utils.VerificationModeUtils.once;
 
 @RunWith(MockitoJUnitRunner.class)
 public class MediationSuccessfulCallbackHandlerTest {
@@ -47,12 +52,14 @@ public class MediationSuccessfulCallbackHandlerTest {
 
     @Mock
     private NotificationTemplates notificationTemplates;
+    @Mock
+    private AppInsights appInsights;
 
     private CallbackParams callbackParams;
     private static final String AUTHORISATION = "Bearer: aaaa";
 
     private Claim claimSetForMediation =
-            SampleClaim.getWithClaimantResponseRejectionForPartAdmissionAndMediation();
+        SampleClaim.getWithClaimantResponseRejectionForPartAdmissionAndMediation();
 
     private CallbackRequest callbackRequest;
 
@@ -65,30 +72,31 @@ public class MediationSuccessfulCallbackHandlerTest {
     public void setUp() {
 
         mediationSuccessfulCallbackHandler = new MediationSuccessfulCallbackHandler(
-                caseDetailsConverter,
-                notificationService,
-                notificationsProperties
-                );
+            caseDetailsConverter,
+            notificationService,
+            notificationsProperties,
+            appInsights
+        );
         when(notificationsProperties.getTemplates()).thenReturn(notificationTemplates);
 
         callbackRequest = CallbackRequest
-                .builder()
-                .caseDetails(CaseDetails.builder().data(Collections.emptyMap()).build())
-                .eventId(MEDIATION_SUCCESSFUL)
-                .build();
+            .builder()
+            .caseDetails(CaseDetails.builder().data(Collections.emptyMap()).build())
+            .eventId(MEDIATION_SUCCESSFUL)
+            .build();
 
         callbackParams = CallbackParams.builder()
-                .type(CallbackType.SUBMITTED)
-                .request(callbackRequest)
-                .params(ImmutableMap.of(CallbackParams.Params.BEARER_TOKEN, AUTHORISATION))
-                .build();
+            .type(CallbackType.SUBMITTED)
+            .request(callbackRequest)
+            .params(ImmutableMap.of(CallbackParams.Params.BEARER_TOKEN, AUTHORISATION))
+            .build();
 
         when(notificationTemplates.getEmail()).thenReturn(emailTemplates);
 
         when(emailTemplates.getClaimantMediationSuccess())
-                .thenReturn(MEDIATION_SUCCESSFUL_CLAIMANT);
+            .thenReturn(MEDIATION_SUCCESSFUL_CLAIMANT);
         when(emailTemplates.getDefendantMediationSuccess())
-                .thenReturn(MEDIATION_SUCCESSFUL_DEFENDANT);
+            .thenReturn(MEDIATION_SUCCESSFUL_DEFENDANT);
 
         when(notificationsProperties.getFrontendBaseUrl()).thenReturn("BASELINE_URL");
     }
@@ -96,40 +104,55 @@ public class MediationSuccessfulCallbackHandlerTest {
     @Test
     public void shouldSendNotificationToDefendant() {
         Claim claim = claimSetForMediation.toBuilder()
-                .response(
-                        SampleResponse
-                                .FullDefence
-                                .builder()
-                                .withMediation(YesNoOption.YES).build()
-                ).build();
+            .response(
+                SampleResponse
+                    .FullDefence
+                    .builder()
+                    .withMediation(YesNoOption.YES).build()
+            ).build();
 
         when(caseDetailsConverter.extractClaim(callbackRequest.getCaseDetails())).thenReturn(claim);
 
         mediationSuccessfulCallbackHandler.handle(callbackParams);
 
         verify(notificationService).sendMail(eq(claim.getDefendantEmail()),
-                eq(MEDIATION_SUCCESSFUL_DEFENDANT),
-                any(),
-                eq("to-defendant-mediation-successful"));
+            eq(MEDIATION_SUCCESSFUL_DEFENDANT),
+            any(),
+            eq("to-defendant-mediation-successful"));
     }
 
     @Test
     public void shouldSendNotificationToClaimant() {
         Claim claim = claimSetForMediation.toBuilder()
-                .response(
-                        SampleResponse
-                                .FullDefence
-                                .builder()
-                                .withMediation(YesNoOption.YES).build()
-                ).build();
+            .response(
+                SampleResponse
+                    .FullDefence
+                    .builder()
+                    .withMediation(YesNoOption.YES).build()
+            ).build();
 
         when(caseDetailsConverter.extractClaim(callbackRequest.getCaseDetails())).thenReturn(claim);
 
         mediationSuccessfulCallbackHandler.handle(callbackParams);
 
         verify(notificationService).sendMail(eq(claim.getSubmitterEmail()),
-                eq(MEDIATION_SUCCESSFUL_CLAIMANT),
-                any(),
-                eq("to-claimant-mediation-successful"));
+            eq(MEDIATION_SUCCESSFUL_CLAIMANT),
+            any(),
+            eq("to-claimant-mediation-successful"));
+    }
+
+    @Test
+    public void shouldRaiseAppInsight() {
+
+        Claim claim = claimSetForMediation.toBuilder()
+            .features(Collections.singletonList(FeaturesUtils.MEDIATION_PILOT))
+            .build();
+
+        when(caseDetailsConverter.extractClaim(any(CaseDetails.class))).thenReturn(claim);
+
+        mediationSuccessfulCallbackHandler.handle(callbackParams);
+
+        verify(appInsights, once()).trackEvent(eq(AppInsightsEvent.MEDIATION_PILOT_SUCCESS),
+            eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
     }
 }
